requirements.txt
django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
phonenumbers==8.12.53
Pillow==9.3.0
djangorestframework-camel-case==1.3.0
python-decouple==3.6
requests==2.28.1
gunicorn==20.1.0
django-grappelli==3.0.4
psycopg2-binary==2.9.3
django-modeladmin-reorder-0.3.1